requirements.txt
torch==1.7.0+cu101
albumentations==0.1.12
pandas==1.1.5
numpy==1.19.5
torchio==0.18.29
tqdm==4.41.1
Pillow==8.1.0
